feat(Interaction): add ability to interact and grab objects
The two scripts that have been added work together by the
`SteamVR_ObjectGrab` script being assigned to a Controller and the
`SteamVR_InteractableObject` script being assigned to any object that
is to be picked up by the controller.

A new example scene has been added which demonstrates the object
interaction and grabbing.