feat(ControllerInteract): add ability to hide controller model
There are two new flags on the Controller Interact script:

  * Hide Controller On Touch
  * Hide Controller On Grab

If they are ticked then the controller model will disappear either
when touching an interactable object or grabbing it (depending on
the flag that is checked).

If Hide on Touch is checked then it will naturally be hidden on
grab without needing to tick the checkbox.