feat(Interaction): add ability to use an interactable object
The Interactable Objects can now also be used as well as grabbed.
A used object just triggers an action to take place on the object
rather than pick the object up and snap it to the controller.

An example scene has been added that demonstrates a usable door by
opening and closing the door when the controller interacts with it.