fix(Teleport): ensure the headset offset is taken into consideration
Previously, when teleporting, the `[CameraRig]` play area centre would
always be where the new teleport location was and it would be jarring
as the headset location within the play area wasn't accounted for.

This fix ensures the headset location within the play area is
compensated so the headset location ends up at the teleported location
and not the actual play area.