fix(Teleport): ensure headset eye object is found
For some reason in versions of Unity (maybe 5.4 beta) the Camera
model underneath the CameraRig prefab gets put at the top level
at run time so the Cemera is no longer a child of the CameraRig.

This causes an issue where the `eyeCamera` cannot be set because
no children exist that have the `SteamVR_GameView` script on them.

As there will only ever be one Game View (i.e. headset) in the
game world then it is an appropriate fix to just find that object
in the entire game world rather than just looking in the children of
the CameraRig prefab.