chore(Interaction): separate the object interaction code
Previously all of the object interactions like touch, grab and use all
existed in the same script (`SteamVR_ControllerInteract`) which meant
that one script was doing a lot and made it very difficult to customise
functionality without growing the script.

Now these three interactions have been separated into their own scripts
although the grab and use scripts require the touch script to also be
applied to the Controller as they use the `touched` object state to
know which object to interact with.

A new Controller Event has also been added to separate grabbing from
using. So by default both interactions use the Controller `Trigger`
but it's now possible to set a different button to grab and to use.