feat(Teleport): add ability to traverse terrain objects
The Basic Teleport now determines the Y position of the play area
based on the terrain height underneath the player. This means that
if the teleport location is on a terrain object then the Y position
of the play area is adjusted to the correct height of the terrain
mesh.

This means it is now possible to traverse terrains using either the
Basic or Height Adjust Teleporters.

A new example scene has been added to show this terrain traversal.