feat(Headset): add headset collision fading
This script is attached to the `Camera (head)` object within the
`[CameraRig]` prefab and adds a box collider and rigidbody to the
headset object. When the headset collides with another collidable
object, the trigger fades the headset view to a given colour.

This is to prevent the player from putting their head into objects
and seeing unwanted object clipping artifacts.

An example scene has been added to demonstrate the new script.