feat(Pointer): add play area cursor option to all pointers
The World Pointer abstract class now generates a play area cursor
that can be projected on to the end of any cursor that extends it.

The play area cursor can also handle collisions on the cursor end and
determine if the play area will be obstructed by other objects within
the cursor location. This can then be used to prevent teleporting
which will prevent players being in locations that they can clip into
objects.

The Simple Pointer and Bezier Pointer have been refactored so their
code that is the same between the two classes now sits in the World
Pointer class (as it's generic to all pointers). This provides the
two child pointers with more default parameters that are generic to
all child pointers as well.

The `Interfaces` directory has been renamed to `Abstractions` as the
World Pointer class isn't actually an interface so made no sense
calling the containing directory `Interfaces`.

A custom shader has been written to allow for Unlit Transparent colours
which work well with the pointers.

An example has been added that shows how the play area with collision
can be used to prevent players from teleporting to locations that would
mean they have the potential to walk into objects.