feat(Interaction): add ability to snap object rotation on grab
The Interactable Object script now has an additional parameter of:

  * `snapToRotation`

This parameter is a `Vector3` that determines the rotation of the
object when it is grabbed. To not set a rotation, the parameter
should be left as `0,0,0` (Vector3.zero).

When a snap rotated object is grabbed, the collisions on the object
are paused for a fraction of a second to prevent colliding with objects
that the grabbed object is on or near.

A new example scene has been created to demonstrate this functionality.