feat(Interaction): add varying grab snap mechanics
Previously, only two snap mechanics were included, this has now been
formalised and an additional mechanic has been added. A new script
parameter has been added to the Interactable Object, the
`Grab Snap Type` determines which grab mechanic an interactable object
will use and can be one of three types:

  * `Simple_Snap`
  * `Rotation_Snap`
  * `Precision_Snap`

The `Simple_Snap` is the original grab snap mechanic where an object
is just centered to the controller position. The `Rotation_Snap` is
the rotational snap mechanism driven by the `Snap To Rotation`
parameter. The new mechanic is the `Precision_Snap` which does not
change the object's rotation or position so it simulates a hand
picking up an object with precision.

The example scene `Examples/014_Controller_SnappingObjectsOnGrab` has
been updated to demonstrate these new snap mechanics.