fix(Pointer): ensure camera rig rotation affects pointer rotation
Previously, if the `[CameraRig]` was rotated then the pointer beam
would come out of the controller at the wrong angle and an additional
AxisType setting was required to correct it. However, this only
corrected the issue for two distinct `[CameraRig]` rotated positions.

This new fix allows the pointers to always project forward regardless
of the rotation of the `[CameraRig]` they even stay in the correct
direction when the `[CameraRig]` is rotated in real time.