fix(Teleport): prevent adjusting height to controller or grabbed object
It was possible using the Height Adjust Teleporter to put the
controller in the path of the raycast that handles the Drop To Floor
check meaning the Play Area would think the controller is a viable
floor and change the Y position to the controller's Y position.

It was also possible to grab an object and use that as something the
ray would hit and drop to an object being held.

These two cases are now resolved by ignoring grabbed objects with
the Drop To Floor ray and the Controllers in the `[CameraRig]` have
been set to the `Ignore Raycast` layer as they should never need
to interact with the world rays.