feat(Pointer): add ability to provide custom play area cursor size
The Play Area Cursor acts as a visual marker of where the play space
will be when the player teleports. It defaults to the actual
calibrated play space if the values are left as zero. If values are
entered then the cursor is displayed in the given size at the
thickness of the border specified on the `SteamVR_PlayArea` script.