feat(Pointer): add option to have beam always on
Sometimes it may be required to have the pointer beam always visible
but still only perform an action on button click. One example could
be when a user is in a menu selection and the beam is always visible
to be used as a pointer and the button press just confirms the
selection.