feat(Teleport): add distance blink delay option
The new Distance Blink Delay option allows for the blink effect on
teleport to be delayed when the user is moving over further distances
to simulate that more time is passing the further the teleport
distance is.

At the moment, the max delay is 1 second that is hard coded because
any longer than this seemed to break immersion. If it is a requirement
to make this a public variable then it can be easily achieved.