feat(Pointer): add rotation to tracer objects in bezier pointer
Previously, the tracer objects in the Bezier Curve would track the
rotation of the controller. This looked bad if objects were not
spherical. However, the objects now rotate to look at the next
object in the beam so they follow the arc of the curve better.

To prevent the objects clipping with the controller on rotation, if
a custom object is used, the first and last tracer object are
disabled.

The float value that provides the curve offset on the beam has now
been pulled out into a public variable called `Beam Curve Ofsset` that
can be used to adjust the arc of the beam if necessary.