feat(Interaction): add events to interactable objects
The Interactable Object now emits events for the three actions that
can be undertaken on the object: Touch, Grab, Use.

When another object touches, grabs, uses (or untouches, ungrabs, stops
using) the Interactable Object then the event is emitted with the
object that is doing the interacting within the payload.