fix(Controllers): add listeners to controllers when they are turned on
Previously, the listeners on the controllers for teleporting and
touchpad walking would only register if the controllers were on and
tracking at the point the game started. If a new controller was
connected (or an existing controller disconnected then reconnected)
the listeners would not be added to that controller.

This fix listens for the `device_connected` event on the controllers
and adds the listener to that controller when it is found. If the
controller is disconnected, the event isn't removed as it will be
valid when the device is reconnected.