feat(Teleporter): add ability to exclude objects as valid locations
It is now possible to tag an object with a specific tag or add a script
to an object with a specific name and specify on the Teleporter script
the name of the tag or script to exclude.

If the pointer destination is touching an excluded object then it will
not be considered as a valid location.