feat(Interaction): UI controls
This introduces UI controls into the framework. They work by
transforming ordinary game objects into interactable elements which
also provide events and values. Some controls feature already an
auto-detect mode that will fully automatically setup the axis inputs.

An abstract base class VRTK_Control provides the basic control flow.

Gizmos are used in the scene to visualize the state of the UI.

The initial set includes the following controls:
* Push button
* Knob
* Slider
* Lever

The lever still contains some work in progress. As this does not
influence productive code it should not be an issue though and
others can expand on the ideas there.

There is a new sample scene included showcasing all controls and the
readme is extended as well.