chore(Interaction): move rigidbody creation from grab to touch
The Interact Touch script is now responsible for creating the
controller rigidbody as it can now also be toggled on or off via a
public method, meaning a grab action isn't required to toggle the
rigidbody state.

The Interact Grab can still toggle the rigidbody, so the exist
methods are unchanged and still work as they previously did.

The rigidbody colliders also are constructed out of boxes to better
fit the shape of the controller rather than a simple sphere.