feat(Teleport): add example of teleporting without using pointers
Now the Destination Events have been separated, it's possible to
trigger a teleport event via a means other than the world pointers.

The example scene shows how a model village of objects can be used
as map markers that determines where the user will be teleported to
upon touching and using one of the markers on the Play Area map.