feat(Interaction): add haptic feedback on object interactions
It's now possible to trigger the controller haptic feedback rumble
when an Interactable Object is touched, grabbed or used.

At present, the inspector parameters shows the label for `duration` as
`x` and the label for `strength` as `y` because the underlying data
structure is a vector2 to hold the required numbers. In the future,
these labels will be replaced when a custom editor is implemented for
the Interactable Object.