feat(Interaction): add ability to get grabbing object
The Interactable Object now has a `GetGrabbingObject` public method
which allows the controller that is holding the object to be obtained
by another script. This is useful if something else in the game world
destroys the held object and a force drop can be initiated on the
grabbing object to reset the controller if required.