feat(Interaction): add additional slider auto-detect mode
So far one could only select to auto-detect axis and min/max together.
If axis detection does not work one could set it manually but then
would also have to define the bounds for min and max. Now it is
possible to specify the axis manually but let the min/max be
auto-detected.

In addition the auto-detection can now also detect setups where the
slider is not necessarily inside a frame but instead on an object.