feat(Headset): add option to ignore certain objects for headset fade
It is now possible to specify a tag or a script name on the
Headset Collision Fade script in the `Ignore Target With Tag Or Class`
parameter. If the object that the headset collides with has the named
tag or contains the named script then the headset will not fade.