chore(Interaction): remove grab snap type parameter
As the `Rotation_Snap` has been removed, it makes little sense to have
a dedicated paramter to determine snap type as this can be inferred by
the two new parameters that have been added.

The `Precision Snap` checkbox determines if a Precision grab snap type
should be used and grabs the object at the point of contact without
adjusting the position or rotation of the object.

The `Right Hand Snap` and `Left Hand Snap` are now used to determine
if the object should be snapped to a specific position and rotation
within the controller. If a `Right Hand Snap` is present then when
the right hand controller picks up the object, this snap point will
be used. The same is true for the left hand controller and the
`Left Hand Snap` point. If neither snap points are provided then
the object is grabbed from the centre (the same as the old
`Simple Snap`).

This cleans up the parameter choices making it easier to understand.

The example scenes have been updated to reflect these changes.