fix(Interaction): ensure touched object is always set after quick grab
There was an issue when grabbing and releasing an object then quickly
grabbing again would leave the touchedObject in a delayed state of
not being touched, meaning the use method on an object would not
trigger because it thought the object being grabbed wasn't being
touched. It would catch up a second or so later, but meant that
objects would feel like they couldn't be used on first pick up.

This commit now resolves the issue by checking on the use method if
the touched object is null then do a sanity check on the grabbed
object.