chore(Interaction): consistent interaction buttons across examples
Previously, the example scenes had different button configurations for
interactions, some scenes used the Grip button for teleport and the
Trigger button for grab, where as other scenes used Touchpad for
teleport and the Grip button for grab.

This made it confusing to know what the buttons were per each scene.
They are now consistent across all examples with the following config:

  * `Touchpad Press` - Pointer
  * `Grip` - Grab
  * `Trigger` - User