feat(Pointer): allow pointers to be placed on any object
The World Pointer scripts can now be placed on any object instead of
just a controller. This enables things like gaze pointers (projecting
a beam from the headset) and works by linking the controller to
the script so there is a way of initiating the destination set event.

The controller linking is done via a new parameter called `controller`
that will auto populate if the script is added to an existing
controller object, but will need populating with a controller object
if the script is attached to another object.

The parameter `beamAlwaysOn` has been replaced with `pointerVisibility`
as this now has three states of:

 * `On_When_Active` only render the pointer beam when the controller
 button is pressed.
 * `Always_On` always render the pointer beam
 * `Always_Off` never render the pointer beam

An example scene has been added to demonstrate attaching pointers to
other game objects such as the headset.