feat(Interaction): add custom controller model example
An example scene has been created that demonstrates how to use a custom
controller model instead of the default HTC Vive controllers. The scene
uses a `Hand` prefab to represent a very basic human hand along with a
`Controller_Hand` script on the `Hand` prefab that does a very basic
animation to denote gripping and using.