fix(Interaction): fix issue with child on controller grab
The grabbed object was not being unparented from the controller
when it was released due to a change which was supposed to fix a bug
that would cause a crash when exiting a scene whilst holding an
object.

The new check doesn't check for the existense of a previous parent
but checks if the object is currently active.