fix(Pointer): correctly plot play area cursor based on offset
Previously, the play area cursor would be plotted based on the headset
location within the play area, but this is not the correct behaviour if
the Teleporter script has the `Headset Position Compensation` parameter
unchecked as the headset position should not be taken into
consideration, otherwise the `[CameraRig]` is teleported to the wrong
destination out by the margin of the headset offset.

This fix ensure the teleport option is honoured in the pointer.