feat(Interaction): add drawer control
This will transform game objects into drawers for use e.g. inside
cabinets. Most things are auto-detected like the direction and the
extent it can be pulled out.

The value range will be 0 (closed) to 100 (fully open). Optionally
interactable objects inside the drawer can become managed so that they
are not grabbable when the drawer is closed.