fix(Interaction): ensure tracked object stays grabbed on teleport
Previously, if a tracked object was being grabbed and the player
teleported, the tracked object would not have it's position updated
after the teleport event, meaning the tracked object would whizz to
the player's current location and detach from their hand.

This fix, listens out for any teleport events and updates the
tracked object's position based on the controller that is holding it.