feat(Pointer): allow exclusions for play area cursor collision
The Ignore Target With Tag Or Class parameter has been added to the
World Pointer script to work alongside the Play Area Cursor collisions
so it's possible to ignore objects from triggering the Play Area Cursor
collision.

This can be useful if objects are on the floor (that may be able to be
picked up) but it's still possible to teleport into their location,
whereas previously, the Play Area Cursor would have collided with them
and prevented teleporting.