fix(Pointer): ensure activate delay only happens on destination set
Previously, the activation delay on a pointer would happen when the
pointer was disabled, but this gave the impression that the pointer
was not working because it was being delayed.

It's better that if no destination is set then the pointer can work
again immediately, but after a destination set has occured then the
pointer activation should be delayed.