feat(Interaction): service pack for 3D controls
This patch is the result of adapting the first set of controls to a
real game scene.

So far only the drawer allowed to set an optional content object. This
was now refactored into the control and now also the door and chest can
profit from the same logic.

In addition objects that were taken out of a drawer would disappear
again if the drawer was closed even if the objects were outside. This
needed to be handled with custom logic. Now there is a helper script
that will manage reparenting automatically.

In addition some code cleanup was done and where already applicable the
direction was extracted with the ultimate goal to have only one
direction enum.

Finally the button contains a small bug fix to define the force in a
much more reliable way.