feat(Interaction): override controller hide + use only if grabbed
Add to Interactable Object a "Hide on touch/grab/use" option which
overrides the controller option (see issue #79 and #346).

Add to Interactable Object a "Use Only If Grabbed" option which can
prevent an object to be used if not grabbed (see issue #346).

Set "Hide Controller On Use" to "Override Hide" and checked
"Use Only If Grabbed" in the object "Usable Grabbable Sphere"
(script "Use Rotate") in the example scene
036_Controller_CustomCompoundPointer. The sphere can be used only if
grabbed and when it is used the grabbing controller disappears.