fix(Controller): check device index is valid in OnDisable
The `OnDisable` method attempts to get the device from the tracked
controller index. However, sometimes during initialisation this
value is set to the max uint value of `4294967295` and causes the
underlying SteamVR code to crash. This fix applies a check to make
sure the index is under the max uint value.