feat(UI): add ability to ignore certain world canvases
A new inspector parameter has been added that provides the ability
to ignore certain world canvases that match a given tag or contain
a specific named class.

The `Ignore Canvas With Tag Or Class` parameter accepts a string and
any canvas that is tagged with the same string or contains a script
that is named the same as the string will not be interactable with
the UI Pointers.

The UI example scene has been updated to demonstrate this exclusion.