fix(Interaction): ensure controller rigid body is created correctly
Due to a bad manual merge into the climbing pull request, the
controller rigidbody was not getting created correctly.

This has now been fixed so the rigidbody is created correctly.

A couple of tweaks to the Docs have been done just changing the word
`you` out of the first person to the third person (e.g. `the user`).