feat(Interaction): add custom controller colliders for touch
It is now possible to provide a custom collection of colliders as a
child object to the Controller. Which means the touch collider is
no longer a simple sphere but can be a complex collection of colliders
to provide a higher level of fidelity.

The `Custom Rigidbody Object` on the Interact Touch script can take
a custom game object that deals with the collider collection for touch
collisions but will also form the collection of colliders for when
the controller rigidbody is not kinematic. If no custom rigidbody is
provided then the default `HTCVive` collider collection will be used.

The example scene `032_Controller_CustomControllerModel` has been
updated to demonstrate how a custom controller can be set up with this
new method.