fix(Controller): ensure rigidbody on not touching works on auto drop
The `Create Rigid Body When Not Touching` flag on the Interact Grab
script would only create a rigidbody on the grab event, so when an
item was being held but then dropped automatically (e.g. by banging it
into a wall and breaking the joint) but the grab button was still being
held down, the rigidbody would not activate.

This fix moves the logic into the Update routine so it constantly
checks for the state of the grab button and toggles the rigidbody
accordingly.