feat(Structure): decouple SteamVR dependency from toolkit
Previously, the SteamVR Unity plugin was a hard dependency as the
scripts for the toolkit made multiple direct calls to SteamVR, OpenVR
and Valve.VR methods. This meant that using an alternative SDK would be
difficult or impossible to use.

This structural change moves all of the calls to SteamVR into an
SDK Bridge script and the toolkit now just makes generic calls to
this SDK Bridge. This means that as long as an SDK has a supporting
SDK Bridge script that supports the required methods then it can be
dropped into the toolkit and should just work.