feat(Controller): add trigger click threshold parameter
It's now possible to specify the click threshold that the trigger axis
must exceed for a click to be registered.

The reason for this is, the trigger click should happen when the axis
has reached it's maximum level of 1f but this is achieved by a
microswitch inside the controller trigger being pressed and when it is
pressed it changes the trigger axis to 1f. However, if the controller
microswitch is broken and not registering the click then the trigger
axis is never set to 1f and therefore the trigger click event is never
emitted. This setting gives the ability to reduce the threshold to a
level where a click even can still be recognised on faulty hardware.