feat(Interaction): add stay grabbed on teleport option
A new `Stay Grabbed On Teleport` option has been added to the
Interactable Object which when checked will ensure the grabbed object
stays with the user when they initiate a teleport. If it is unchecked
then the object will be force dropped when the user teleports. This can
be very useful for interactable objects like door handles, or chest
handles that would normally disjoint if a teleport event occurs whilst
they are being grabbed.