fix(Interaction): make tracked objects throw the same as other objects
Tracked objects were not being thrown in the same way as other object
grab types like joints or child of controller meaning that they would
be thrown with a lot more force which made them feel unnatural. This
has been resolved to ensure they feel the same as other objects when
thrown.