chore(Teleport): remove static events in dash teleport
All of the other scripts use non-static events and require an instance
of an object to register events on it. Whilst items like the dash
teleport may only have one script in the scene, it's still more
appropriate to expect an instance because this way it ensures the
script is available on an object in the scene to register events
against.