chore(Interaction): change variables from private to protected
The Interactable Object is expected to be extended and a number of
protected methods use class variables that are private which causes
issues if these protected methods need to be overriden and extended.

It's more appropriate to make these variables protected as well so
any extended class can read these variables and adjust accordingly.