chore(Structure): move prefab scripts within prefabs directory
The scripts that the prefabs rely on have now been moved into a
scripts directory within the prefabs directory as it makes more
sense to group these items together rather than keep the scripts
in the helper folder.

The radial menu scripts are still contained within the radial menu
directory in the `scripts/2D` directory.