chore(Structure): add namespace to radial menu prefab scripts
The radial menu scripts are decoupled from VRTK however, to prevent
clashes with other classes of the same name, it's easier to just
put these classes in to the `VRTK` namespace as they are required for
VRTK and not really envisaged to be used independently of the toolkit.

The Interactable Object Editor has also been moved to the `VRTK`
namespace for completeness.