fix(Interaction): ensure undroppable objects cannot be dropped
If the Interactable Object is set as `isDroppable = false` then it
should not be able to drop the object by the joint breaking or
the object/controller distance being too far. It should also not
drop if the controller is disabled but it will drop if the
Interactable Object is disabled.

The way the prevent drop works on controller disable is by storing
a reference to the grabbed object then auto grabbing it again
when the controller is re-enabled.