chore(DeviceFinder): update scripts to use DeviceFinder over SDK Bridge
The `GetControllerRightHand` and `GetControllerLeftHand` methods now
also exist in the Device Finder script as they are devices that are
being found. They still just proxy the call to the SDK Bridge script
but it's more appropriate to use the highest access point possible
rather than going straight to the bridge if needed.