chore(Pointer): improve pointer material transparency
The existing World Pointer material using the new shader was set so
it was too transparent by default even if the pointer colour was
set to 100% opaque. The material has been tweaked so it looks more
opaque and the transparency setting on the pointer colour can be
used to reduce the opacity if required.