fix(Controller): ensure controller elements are cached initially
The controller elements that are used for highlighting are
cached but they can initially be set to null but still exist in
the cache. This causes highlighting to break until the scene
is reloaded. The fix checks to see if the cached element is null
and if it is then it sets the cache until it's not null.