feat(PolicyList): add ability to have multiple tag/script ignore checks
Previously, many of the classes utilised a single string to determine
whether something should be ignored by a tag or a script on the
component.

For example, the Teleport scripts could have a single tag or class name
provided to ignore any game objects containing that tag or named
component and treat it as an invalid teleport location.

This was very limiting as it only allowed for a single tag or class to
be specified and was always based on the items had to be in the ignore
list to be limited.

The new Tag Or Script Policy List script can now be added to a
component and have multiple tags or script name elements checked
against and can also have the operation set to `Ignore` to provide
the current behaviour, or be set to `Include` to only allow operations
on identifier matches.

This means for instance that the teleport script can be set up to
only teleport to areas that have a tag or script applied that is
in the list (rather than building a list of areas that cannot be
teleported to).

The existing Ignore With Tag Or Class string field has been left in
for legacy purposes and in case there are any unforseeable issues
with this new policy list. But it should be deprecated and then
removed in the future.