feat(UIPointer): add option to do click when pointer is deactivated
A new option on the UI Pointer script is now available that when
enabled will automatically attempt a click event on the last hovered
UI element before the UI Pointer was deactivated.

This provides the ability to just enable the UI Pointer and point at
an object then release the pointer to perform a click, rather than
having to press a separate button to perform the click action.